requirements.txt
itchat==1.3.10
google-generativeai==1.4.0
